Port to DX12 and cleanup.

+++ pcsx2/GS/Renderers/DX12/GSDevice12.h
@@ -353,7 +353,7 @@ private:
 	ComPtr<ID3D12PipelineState> m_shadeboost_pipeline;
 	ComPtr<ID3D12PipelineState> m_imgui_pipeline;
 
-	std::unordered_map<u32, ComPtr<ID3DBlob>> m_tfx_vertex_shaders;
+	std::unordered_map<u64, ComPtr<ID3DBlob>> m_tfx_vertex_shaders;
 	std::unordered_map<GSHWDrawConfig::PSSelector, ComPtr<ID3DBlob>, GSHWDrawConfig::PSSelectorHash>
 		m_tfx_pixel_shaders;
 	std::unordered_map<PipelineSelector, ComPtr<ID3D12PipelineState>, PipelineSelectorHash> m_tfx_pipelines;
